Update multigrid.h
Update with inexact Newton method descried in that paper. Can converge to 1e-5 for 64*64*64 grids. The converging speed is slow because I have not found a quick way to let the corresponding Jacob linear system converge. The parameter C and p control the converging speed which need to be fine tuned when dealing with specific problem I think.